requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.13.1
Flask==3.1.0
flask-cors==5.0.1
flask-talisman==1.1.0
fonttools==4.57.0
fsspec==2024.6.1
gevent==25.5.1
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
mpstemmer @ git+https://github.com/ariaghora/mpstemmer.git@25a5fd923af163a7eac3a5ec976984156ca8fa8b
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
ollama==0.4.7
openai==1.66.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
preshed==3.0.9
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
Sastrawi==1.0.1
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
sympy==1.13.3
thinc==8.3.6
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
